pyproject.toml

[project]
name = "snooker-score-bot"
version = "0.1"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.115.6",
    "google-cloud-logging==3.11.3",
    "gspread==6.1.4",
    "httpx==0.28.1",
    "jinja2==3.1.5",
    "langchain-google-vertexai==2.0.21",
    "langchain==0.3.24",
    "orjson==3.10.14",
    "pydantic==2.9.2",
    "python-multipart==0.0.20",
    "pytz==2024.2",
    "tabulate==0.9.0",
    "twilio==9.4.1",
    "uvicorn==0.34.0",
    "click",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/snooker_score_bot"]

[project.optional-dependencies]
dev = ["python-dotenv", "pytest", "pytest-dotenv", "ruff", "freezegun"]

[tool.ruff]
line-length = 119

[tool.ruff.lint]
select = ["E", "F", "I"]
